requirements.txt
mysql-connector-python
pandas
jupyter
